Revert "feat: add campaign update endpoint with blockchain integration and audit trail"
