+++ src/components/v1/Donation/donation.dto.ts
@@ -1,0 +1,14 @@
+export interface CampaignResponseDto {
+  campaignId: string
+  campaignRef: string
+  targetAmount: string
+  donationToken: string
+  transactionHash: string
+  createdAt: Date
+}
+
+export interface ApiResponse<T> {
+  data: T | null
+  success: boolean
+  message?: string
+}
