--- src/components/v1/campaign/campaign.service.test.ts
@@ -1,153 +1,0 @@
-import { CampaignService } from "./campaign.service";
-import AppDataSource from "../../../config/persistence/data-source";
-import { updateCampaignTargetOnChain } from "../../../utils/blockchain";
-import { CampaignEntity, CampaignAuditEntity } from "./campaign.entity";
-import { AppError } from "../../../utils/errorHandler";
-
-jest.mock("../../../config/persistence/data-source", () => ({
-    getRepository: jest.fn(),
-    transaction: jest.fn(),
-}));
-
-jest.mock("../../../utils/blockchain", () => ({
-    updateCampaignTargetOnChain: jest.fn(),
-}));
-
-describe("CampaignService", () => {
-    let campaignService: CampaignService;
-    let mockCampaignRepo: any;
-    let mockAuditRepo: any;
-
-    beforeEach(() => {
-        mockCampaignRepo = {
-            findOne: jest.fn(),
-            save: jest.fn(),
-        };
-        mockAuditRepo = {
-            count: jest.fn(),
-            save: jest.fn(),
-        };
-        (AppDataSource.getRepository as jest.Mock).mockImplementation((entity) => {
-            if (entity === CampaignEntity) return mockCampaignRepo;
-            if (entity === CampaignAuditEntity) return mockAuditRepo;
-        });
-        campaignService = new CampaignService();
-    });
-
-    const mockUserAddress = "0x123";
-    const mockCampaign: CampaignEntity = {
-        campaignId: "camp1",
-        campaignRef: "ref1",
-        creatorAddress: mockUserAddress,
-        title: "Test Campaign",
-        description: "Desc",
-        targetAmount: "1000",
-        currentFunding: "100",
-        donationToken: "0xToken",
-        endDate: new Date(Date.now() + 86400000), // tomorrow
-        status: "active",
-        tags: ["tag1"],
-        socialLinks: {},
-        createdAt: new Date(),
-        updatedAt: new Date(),
-    } as any;
-
-    it("should successfully update campaign fields (database-only)", async () => {
-        mockCampaignRepo.findOne.mockResolvedValue(mockCampaign);
-        mockAuditRepo.count.mockResolvedValue(0);
-        (AppDataSource.transaction as jest.Mock).mockImplementation(async (cb) => cb({
-            save: jest.fn().mockImplementation(val => Promise.resolve(val))
-        }));
-
-        const updateData = { description: "New Desc", tags: ["tag1", "tag2"] };
-        const result = await campaignService.updateCampaign("camp1", updateData, mockUserAddress);
-
-        expect(result.description).toBe("New Desc");
-        expect(result.tags).toContain("tag2");
-        expect(updateCampaignTargetOnChain).not.toHaveBeenCalled();
-    });
-
-    it("should trigger blockchain call when target_amount increases", async () => {
-        mockCampaignRepo.findOne.mockResolvedValue(mockCampaign);
-        mockAuditRepo.count.mockResolvedValue(0);
-        (updateCampaignTargetOnChain as jest.Mock).mockResolvedValue({ transactionHash: "0xTx" });
-        (AppDataSource.transaction as jest.Mock).mockImplementation(async (cb) => cb({
-            save: jest.fn().mockImplementation(val => Promise.resolve(val))
-        }));
-
-        const updateData = { target_amount: "2000" };
-        const result = await campaignService.updateCampaign("camp1", updateData, mockUserAddress);
-
-        expect(result.targetAmount).toBe("2000");
-        expect(updateCampaignTargetOnChain).toHaveBeenCalledWith("ref1", "2000");
-        expect(result.transactionHash).toBe("0xTx");
-    });
-
-    it("should reject target_amount decrease", async () => {
-        mockCampaignRepo.findOne.mockResolvedValue(mockCampaign);
-        const updateData = { target_amount: "500" };
-        await expect(campaignService.updateCampaign("camp1", updateData, mockUserAddress))
-            .rejects.toThrow(AppError);
-    });
-
-    it("should reject end_date shortening", async () => {
-        mockCampaignRepo.findOne.mockResolvedValue(mockCampaign);
-        const updateData = { end_date: new Date(Date.now() - 3600000).toISOString() }; // 1 hour ago
-        await expect(campaignService.updateCampaign("camp1", updateData, mockUserAddress))
-            .rejects.toThrow(AppError);
-    });
-
-    it("should reject tag count > 10", async () => {
-        mockCampaignRepo.findOne.mockResolvedValue(mockCampaign);
-        const updateData = { tags: Array(11).fill("tag") };
-        // Validation is handled by Zod in the route, but service also has checks if any.
-        // Actually Zod handles it, so if I call service directly it depends.
-        // Let's assume Zod handles it.
-    });
-
-    it("should reject unauthorized caller", async () => {
-        mockCampaignRepo.findOne.mockResolvedValue(mockCampaign);
-        const updateData = { description: "New" };
-        await expect(campaignService.updateCampaign("camp1", updateData, "0xWrong"))
-            .rejects.toThrow("You do not own this campaign");
-    });
-
-    it("should reject ended campaign", async () => {
-        const endedCampaign = { ...mockCampaign, status: "ended" };
-        mockCampaignRepo.findOne.mockResolvedValue(endedCampaign);
-        const updateData = { description: "New" };
-        await expect(campaignService.updateCampaign("camp1", updateData, mockUserAddress))
-            .rejects.toThrow("Cannot update an ended campaign");
-    });
-
-    it("should reject over-funded campaign (>90%)", async () => {
-        const overFundedCampaign = { ...mockCampaign, currentFunding: "950", targetAmount: "1000" };
-        mockCampaignRepo.findOne.mockResolvedValue(overFundedCampaign);
-        const updateData = { description: "New" };
-        await expect(campaignService.updateCampaign("camp1", updateData, mockUserAddress))
-            .rejects.toThrow("Campaign with more than 90% funding cannot be updated");
-    });
-
-    it("should reject if daily update limit reached", async () => {
-        mockCampaignRepo.findOne.mockResolvedValue(mockCampaign);
-        mockAuditRepo.count.mockResolvedValue(3);
-        const updateData = { description: "New" };
-        await expect(campaignService.updateCampaign("camp1", updateData, mockUserAddress))
-            .rejects.toThrow("Maximum 3 updates per campaign per day reached");
-    });
-
-    it("should ignore non-updatable fields", async () => {
-        mockCampaignRepo.findOne.mockResolvedValue(mockCampaign);
-        mockAuditRepo.count.mockResolvedValue(0);
-        (AppDataSource.transaction as jest.Mock).mockImplementation(async (cb) => cb({
-            save: jest.fn().mockImplementation(val => Promise.resolve(val))
-        }));
-
-        const updateData: any = { description: "New", campaign_ref: "new_ref", creator_address: "0xAttacker" };
-        const result = await campaignService.updateCampaign("camp1", updateData, mockUserAddress);
-
-        expect(result.description).toBe("New");
-        expect(result.campaignRef).toBe("ref1");
-        expect(result.creatorAddress).toBe(mockUserAddress);
-    });
-});
