+++ src/types/global.ts
@@ -28,7 +28,6 @@ export interface IToken {
     sessionId: string;
     ref: number;
     role: string;
-    address: string;
 }
 
 export interface GenerateTokenArg {
