--- src/utils/blockchain.ts
@@ -1,15 +1,0 @@
-/**
- * Mocking Cairo contract interaction for update_campaign_target.
- * In a real scenario, this would use starknet.js or a similar library.
- */
-export const updateCampaignTargetOnChain = async (
-    campaignRef: string,
-    newTargetAmount: string
-): Promise<{ transactionHash: string }> => {
-    // Simulate network delay
-    await new Promise(resolve => setTimeout(resolve, 1000));
-
-    // Simulate success
-    const txHash = `0x${Math.random().toString(16).slice(2, 66).padStart(64, '0')}`;
-    return { transactionHash: txHash };
-};
